libvirt: Add Terraform variables for memory/CPU, bump master to 4GiB
My main dev environment is a Lenovo P50 with 64GB of RAM - I got
it specifically to run some large VMs (and/or many VMs) specifically
with OpenShift in mind.

First, default masters to 4096 MiB since we are seeing a default
install be overloaded.

And for me, increasing RAM on my master to 8GB is a *very* noticeable speed improvement
and I think reliabilty; before I saw the apiserver be `OOMKilled`
sometimes, and `kswapd0` was constantly doing writeback.

These variables aren't bubbled all the way up to the documented
installer config, but one can now do e.g.:

```
$ env TF_VAR_libvirt_master_memory=8192 TF_VAR_libvirt_master_vcpu=4 ./bin/openshift-install create cluster --dir=osiris
```

Previously:

 - https://github.com/openshift/installer/pull/408
 - https://github.com/openshift/installer/pull/163