cmd/openshift-install: RunE -> logrus.Fatal
Before this commit, we had a number of callers with the following
pattern:

  func doSomething(args) error {
    cleanup, err := setupFileHook(rootOpts.dir)
    if err != nil {
      return errors.Wrap(err, "failed to setup logging hook")
    }
    defer cleanup()

    err := someHelper()
    if err != nil {
      return err
    }

    return nil
  }

The problem with that approach is than any errors returned by
doSomething will not be logged to .openshift_install.log, and we
definitely want those errors logged to the file.  With this commit,
I've shuffled things around to ensure we call logrus.Fatal(err) before
the deferred cleanup() fires.