*: include terraform in the installer binary
Instead of requiring that the user has Terraform installed, this now
includes the pre-built Terraform binary within the installer. At
runtime, the installer unpacks Terraform into the temporary directory
and executes it from that location. This also simplifies the version
output - since the Terraform version is known at build-time, there is
no need to show the user. Lastly, this cleans up some of the log
messages so that they don't explicitly refer to Terraform.

Including Terraform in the installer adds 18MB to the binary size,
bringing the total up to 82MB.