types/libvirt: remove per-machine-pool configuration
None of the options, other than Image, that were in the Libvirt MachinePool were being used.
They have been removed. The Image has been pulled up to the Libvirt Platform, as there was no
way to use a different image for different machines pools.

For consistency with the AWS and OpenStack platforms, the Libvirt MachinePool has been retained,
even though it is empty. The DefaultMachinePlatform has been retained in the Libvirt Platform
as well.

The code in the Master Machines and Worker Machines assets that determines the configuration
to use for the machines has been adjusted for Libvirt to rectify the machine-pool-specific
configuration against the default machine-pool configuration. This is not strictly necessary
as, again, the Libvirt configuration is empty. It keeps the logic consistent with the other
platforms, though.

https://jira.coreos.com/browse/CORS-911