*: mirror tectonicClusterID as openshiftClusterID
As part of removing references to tectonic, the tectonicClusterID tags will
be changed to openshiftClusterID. This is the first part of that. This
creates the openshiftClusterID tag that will be laid down alongside the
tectonicClusterID tag. After all of the usages of tectonicClusterID in other
repos are removed, then the tectonicClusterID will be removed, leaving only
the openshiftClusterID.

https://jira.coreos.com/browse/CORS-878