Refactor RHCOS fetch API, support fetching qemu from partial config
The *main* thing I'm really fixing here is that I have an installer
wrapper script: https://github.com/cgwalters/homegit/blob/master/bin/okdinst

And I noticed recently that I kept fetching old RHCOS versions; that's
of course because the image URL was hardcoded in my install config.  What
I really want is to fetch that dynamically.

So this PR changes the TF rendering code to do an on-demand fetch
if the image isn't filled out in the libvirt config.

Now further, as I was arguing in this PR:
https://github.com/openshift/installer/pull/710#issuecomment-440486784

The result of an install run from installer git master is currently a function
of the 3-tuple:
(installer, RHCOS, release payload)

I'd actually like to print out all 3 clearly in the logs; this is
a first step.

Today as far as I can see we do separate fetches in AWS for master,
bootstrap and worker, which is a bit broken.  Need to move that
to a higher level, and this will help with that.