asset/manifests: network and ingress equal from state file and on-disk
The on-disk and in-state-file versions of the Network and Ingress manifest
assets are not evaluating as equal when read into the store. Consequently,
they are being marked as dirty, even when the user has not made any changes
to them. This propagates to the Operators and Openshift manifest assets,
making them dirty as well.

For both assets, the underlying issue is that the config field is not stored
in the state file but it is populated when loading from on-disk. The config
field in manifests.Network has been made public, allowing it to be stored in
the state file. The config field in manifests.Ingress has been removed, since
it is not being used.

Fixes https://jira.coreos.com/browse/CORS-938