asset/manifest: do not load Common Manifest dependent manifests from disk
The network, ingress, and dns manifest assets are not targeted assets.
The installer does not write their files to disk. The installer should
not load their files from disk.

In addition to the load issue, there is an issue with reading the
config field when reading the assets from the state file. For the network
asset, the config field is needed, but is not stored in the state file.
To resolve this, the config field has been made public. For the ingress
and dns assets, the config field is not needed: it has been removed.

Fixes https://jira.coreos.com/browse/CORS-938