assets: move asset store impl to own package
Tests are being created that use the asset store implementation to
validate restoring assets from the state file and disk. The tests are
much easier when they have access to the private fields of the asset
store. Rather than exposing more public fields and methods on the store,
the tests will be placed in the same package as the store. Unfortunately,
this creates cyclic imports between the base asset package and the
various packages containing the targeted assets. To rectify this,
the concrete asset store implementation has been moved to a new
pkg/asset/store package.

Changes for https://jira.coreos.com/browse/CORS-940