assets/templates: remove common templates asset
The Common Templates asset is not truly an asset. It does not own any
on-disk files, and no other assets depend upon it.

This change is part of work that is on-going to add tests to ensure that
targeted assets are not assessed as dirty when read back in without
being modified. The problem with the Common Templates asset is that it
writes files to disk as a targeted asset but then does not load any
assets back in. So a test that verifies that the loaded-in asset matches
the written asset fails.