assets/templates: remove common templates asset
The Common Templates asset is not truly an asset. It does not own any
on-disk files, and no other assets depend upon it. The problem with
the Common Templates asset is that it writes files to disk as a targeted
asset but then does not load any assets back in. So a test that verifies
that the loaded-in asset matches the written asset fails.

Changes for https://jira.coreos.com/browse/CORS-940