pkg/asset/installconfig/pullsecret: Point to cloud.openshift.com
try.openshift.com used to be an HTML redirect to cloud.openshift.com,
but now it's a page in its own right talking about what OpenShift 4
is.  Folks who are trying to find a pull secret for the installer are
already pretty interested, so they shouldn't have to dig too hard to
get the JSON they need.

This will also help avoid confusion like we saw for the CoreOS flow
[1], where the pull secret was not immediately obvious to several
users due to an undocumented "register for a Tectonic plan"
intermediate [2,3].

Currently the JavaScript on /clusters/install is stripping the
'#pull-secret' fragment.  Hopefully that will get sorted out soon and
this link will drop users right into the section with the pull-secret.

[1]: https://github.com/openshift/installer/pull/663#discussion_r234345693
[2]: https://github.com/openshift/installer/pull/677
[3]: https://github.com/openshift/installer/pull/691