asset/templates: remove fileName field
The fileName field is not needed. The template assets do not need to
store this information.

This change is part of work that is on-going to add tests to ensure that
targeted assets are not assessed as dirty when read back in without being
modified. The problem with the fileName field is that it is not set when
the asset is loaded from the state file or from on-disk. This causes
road-trip tests that compare a generated asset to a read asset to fail.