asset/templates: remove fileName field
The fileName field is not needed. The template assets do not need to
store this information. The problem with the fileName field is that
it is not set when the asset is loaded from the state file or from
on-disk. This causes round-trip tests that compare a generated asset
to a read asset to fail.

Changes for https://jira.coreos.com/browse/CORS-940