pkg/asset/installconfig/pullsecret: Use survey.Password
Folks are going to be pasting this in, so they won't have to see the
characters to fix typos and such.  And using the Password input allows
folks to demo the wizard (or run it where they are concerned about
shoulder surfing) and not have to worry about leaking their secret.
It also reinforces the fact that this *is* a secret, which may help
users remember not to post it in public places (e.g. issue reports).