openstack: Add a temporary service VM.
This patch introduces a temporary service VM, that provides enough DNS and LB
functionality to bring up the cluster. It's a VM running dnsmasq to provide the
required DNS records necessary to bring up the cluster and haproxy to LB
against the master nodes for the API.

OpenStack support will remain experimental as long as this VM exists, but it
provides a better test and dev experience while we work on improving.

This includes the folowing additional cleanups in passing:
 - drop unused console security group
 - hack to disable openstack cloud provider temporarily while some issues
   are resolved

Co-authored-by: Flavio Percoco <flavio@redhat.com>
Co-authored-by: Tomas Sedovic <tsedovic@redhat.com>