.github/ISSUE_TEMPLATE: Explicitly mention .openshift_install.log
We're still getting issue reports about installer crashes that attempt
to summarize the issue but do not include copies of the actual error
messages.  While summaries can be accurate, they're usually not much
more compact than a copy/paste of the errors from the logs, and it's a
lot easier to find existing reports for a given error if those reports
include the error you're looking for.