Revert "OpenStack Destroy, tolerate temporary router delete failure"
As reported in #891 this is not the right fix and can result in a
loop if the router delete fails.  Instead we need to remove interfaces
from the router before attempting deletion.  For now revert the
incorrect fix and I'll push a follow-up with revised solution
pending further local testing.

Related: #891

This reverts commit fb73b45ca90d9ffe5d2428841444580a039c2614.