pkg/destroy/aws: Bring over hive/contrib/pkg/awstagdeprovision
We've been meaning to take this off the hands of the Hive folks for a
while.  Finally copy it over (I'll drop the vendored copy soon).  This
is a verbatim copy of the file as it stands with
openshift/hive@ad6f8d5b (Merge pull request openshift/hive#143 from
abutcher/capischeme, 2018-12-18).