pkg/destroy/aws: Pick a shorter package name
The fact that it's a subpackage of pkg/destroy is sufficient context
without repeating "deprovision" in the package name.  And the fact
that the deprovision is tag based is an implementation detail that
doesn't need to be surfaced in the package name.

Also drop the copyright header.  We're also an Apache-2.0 project, so
this just removes the copyright holder information from the header.
And "the Kubernetes Authors" wasn't all that helpful anyway.  If folks
want to see who authored the contents, they should check the Git
history.  I'm breaking that history here, so here's a list of authors
in case the hive repository goes away or something: Abhinav Dahiya,
Dan Mace, Devan Goodwin, Joel Diaz, Miciah Masters, Thomas Wiest,
and me.