*: add support for platform none
This adds support for the "none" platform. This is used in situations
where the infrastructure will not be managed (bring-your-own).
Additionally, since the infrastructure is unmanaged, the installer
cannot create the infrastructure and the "cluster" target is therefore
invalid when used with the "none" platform. For this reason, the "none"
platform is hidden in the UX. It is still possible to use, but the
install-config will need an empty "none" object under platform.