update for docs/user/troubleshooting: Document 'core' user and SSH key injection
Previously, these platform-agnostic docs left users guessing about which username to use when SSHing in.

And folks with AWS experience might be surprised that we don't use AWS key pairs, so add some wording to avoid them thinking the lack of an AWS key pair is the source of their SSH issues.