openstack: document DNS prerequisites
This patch documents the kind of records that must be configured in the
the environment DNS prior to launching a deployment in order for the
etcd discovery and aliases to work properly. This is only for in-cluster
DNS resolution and as such does not include OCP router wildcard record
examples.

Signed-off-by: Antoni Segura Puimedon <celebdor@gmail.com>