data/aws: Tag IAM roles
As reported by Brian Harrington, we're currently deleting these roles
based on cluster name, when we'd ideally be deleting them based on the
more-specific cluster ID [1].  Tagging the roles (new in 2018-11-16
[2]) is a step in that direction, although as of this commit we still
delete roles by name.  In coming work, I'll pivot to deleting these
based on their tags.

The tag property is documented in [3].  Unfortunately, instance
profiles are not tag-able [4].

[1]: https://github.com/openshift/installer/issues/1000
[2]: https://aws.amazon.com/blogs/security/add-tags-to-manage-your-aws-iam-users-and-roles/
[3]: https://www.terraform.io/docs/providers/aws/r/iam_role.html#tags
[4]: https://www.terraform.io/docs/providers/aws/r/iam_instance_profile.html