Delete DNS server from service VM and documentation update
For the Openstack DNS (short-term) solution, the external DNS should be
configured with the relevant SRV and CNAME records.
This patch deletes the DNS server runs in service VM and
documents the kind of records that must be configured in the
the environment DNS prior to launching a deployment in order for the
etcd discovery and aliases to work properly. This is only for in-cluster
DNS resolution and as such does not include OCP router wildcard record
examples.

Co-authored-by: Antoni Segura Puimedon <celebdor@gmail.com>