openstack: one subnet for masters and workers
Moving both masters and workers to the same subnet will simplify the
deployment. It is also useful for the Kuryr trunk vif driver which
expects only a single subnet id and needs to provide for both masters
and workers.

One Subnet to rule them all, One Subnet to find them
One Subnet to bring them all, and in L3 bind them"

Signed-off-by: Antoni Segura Puimedon <celebdor@gmail.com>