destroy/libvirt: cannot destroy domain if it is already shutoff
https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainDestroy
```
Destroy the domain object. The running instance is shutdown if not down already and all resources used by it are given back to the hypervisor.
```

Calling destroy on a shutoff/shutdown domain returns `Requested operation is not valid: domain is not running` error.
Therefore, skipping calling destroy when domain is already in those 2 states.