asset/ignition: add support for specifying owner
This gives the ability to specify the owner of the files. So far, only
the root user is used, but support for other users will be needed in
order to support journal-gatewayd. This service doesn't run as root, so
in order to read its TLS assets, they either need to be world-readable
or they need to be owned by the correct user.