types: hide OpenStack in the list of platforms
Since OpenStack isn't going to be fully supported in 4.0, it needs to be
hidden to prevent users from accidentally choosing it. This still leaves
the platform prompt in the UI with just a single choice, but this is
nice because it makes it clear that more platforms will be supported by
this tool in the future.