Revert "openstack: Del DNS from service VM update docs"
This reverts commit ae54392948cccff9b33abf49522547b3746127f8.

The proposed solution (including the exact same CoreDNS configuration)
does not work for several of our developers deployments causing a
regression (i.e. we're not able to get the etcd cluster started).

It appears that the DNS requests are getting the CNAME values for the
etcd records, but they never resolve to an IP address.

We've not figured out why that is yet, so reverting this change lets us
unblock the OpenStack deployments.