gcp: Enable a GCP installer
This wires a minimum of create logic and behavior to enable a new
platform 'google' (name should be 'gcp', but terraform fails when
platform name != terraform-cloud-provider name). The destroy function
is wired directly to terraform for now and is considered a bare
minimum step.

The terraform steps are consistent with AWS and the 3.11 provisioning
playbooks in openshift-ansible - roughly:

1. Create bootstrap in zone[1], masters in zones[1..3]
2. Use internal network load balancer for masters
3. Use TCP proxy for external load balancer
4. (no domain name setup yet)
5. Directly inject bootstrap.ign to user-data since we can on GCP
6. Using a hardcoded maipo image since we don't have a continuous run
7. Create a service account for the bootstrap
8. Wire through instance size, disk, and disk size