create: separate bootstrap and node ignition config assets into separate targets
The Bootstrap Ignition asset is dependent upon the Master and Worker Ignition assets.
It is problematic to have them all targeted by the same sub-command because it makes
it difficult for a user to modify all of the ignition configs. These changes create
a new sub-command that targets just the Master and Worker Ignition assets.

The ignition-configs sub-command has been renamed to pre-cluster. This reflects that
the sub-command targets more than just the Bootstrap Ignition asset, as it targets
the kubeconfig and the metadata.json as well.

The new sub-command has been named node-config.

The ignitions-configs sub-command has been deprecated and will be rmeoved once it is
no longer used.

Fixes https://jira.coreos.com/browse/CORS-948