data/aws: use azs for master set in manifests
These changes pass the availability zones to use for the masters
set in 99_openshift-cluster-api_master-machines.yaml through to
terraform. Prior to these changes the masters were always placed in
the first 3 availability zones.

There is no validation done on the availability zones to verify that
they are valid for the region.

Fix for https://bugzilla.redhat.com/show_bug.cgi?id=1662119.