data/aws/variables-aws: Drop aws_master_ec2_type and similar defaults
Since 9a9cd6de (pkg/tfvars: Respect install-config AWS machine pools,
2019-01-15, #1076) and cbad1a41 (Add size and type support for aws
volumes, 2019-01-16, #1079), we should always be providing these in
the Terraform variables generated by the asset graph.

There's still a default for the bootstrap machine, but we plan on
removing that from Terraform shortly.

Dropping omitempty for the master IOPS allows the zero value (if IOPS
is nil or zero in the machine config) to generate a non-empty
Terraform value, avoiding:

  Error: Required variable not set: aws_master_root_volume_iops"

I've left omitempty on the other AWS properties, because in most cases
their zero values will not produce a functional cluster and we want to
fail fast.  In the user-tags case, we have a sane (empty set) default,
so omitting the nil map is fine.