pkg/types/aws/machinepool: Drop IAM-role overrides
We're planning on dropping instance profiles in favor of the new
credentials operator [1], because we want AWS access to have
operator/pod/namespace granularity and not instance granularity.  Many
pods could be running on a given instance, and not all of them should
have a given permission.  While we're blocked from dropping these at
the moment due to kubelet cloud-config+secrets [2], we can drop the
user-facing knobs for this feature now.  Then pivoting the internal
approach once we get the kubelet sorted will be a non-breaking change.

[1]: https://github.com/openshift/cloud-credential-operator
[2]: https://github.com/openshift/installer/pull/697#issuecomment-451582731