pkg/asset: Marshal and compare strings instead of DeepEqual
reflect.DeepEqual between the stateFileAsset and onDiskAsset did not work as intended
because onDiskAsset used the 'Load' function which populates some internal private fields also (see dns asset).
Though possibly inefficient, we have to unfortunately marshal the assets back to strings to be able to check
equality.
reflect.DeepEqual was always returning not matched for dns asset as an example, even if the on disk files stay untouched.