pkg/asset: A new asset called user-manifests that just picks up user defined manifests
A new asset called user-manifest is created so that any files that the user wants to get created
can be put in a directory and the installer will end up picking it.
The iginition asset will depend on this asset which will have 'empty' files by default, unless
something is found on-disk which then gets loaded.
It helps us to separate manifests asset and user-defined manifests because there will be no
entanglement with other dirty assets and the user can always just plant the manifests without
running a separate target 'create manifests' as a sub-step to getting stuff added.