OpenStack: switch to CoreDNS
This replaces the dnsmasq container used the OpenStack provider with
CoreDNS. OpenShift already uses CoreDNS so it doesn't make sense to
bring in a different DNS tech to the mix.

This also reads /etc/resolv.conf for upstream DNS rather than
hardcoding OpenStack's internal IP address. And it uses CNAMEs to
provide etcd IPs instead of generating a hosts file.