pkg/asset/machines/aws: Only return available zones
Zones can also be impaired, unavailable, or "information" [1,2].  I
dunno what that last one is about.  But we only want to put machines
in available zones ;).

This is a bit racy, because we could get different responses at both
call-sites.  Moving forward, we'll want to shift this request into a
single available-zones asset.

[1]: https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeAvailabilityZones.html
[2]: https://www.terraform.io/docs/providers/aws/d/availability_zones.html#state