pkg/asset/machines/aws: Cache zone responses
To protect us from allocating masters based on one zone request and
workers based on a different zone request.  Once we get the zone
information from AWS, using the same response in both locations is
more efficient and guards us against divergent responses.  If a zone
goes off- or online while we're working through asset generation,
we'll get a more predicatable topology if we don't alter our zone pool
along the way.  If the change is a zone going offline, the cluster is
probably going to fail.  But it would have failed anyway.  If the
change is a zone coming online, this commit ensures we don't strand
some workers (or masters, depending on which asset was rendered last)
off in the new zone by themselves.  Having stranded workers or masters
would probably work, but might surprise customers who asked for, say,
three of each type of machine.

I prefer having the cached value live in an asset, because users could
configure their desired zones by setting that asset to avoid the need
for an AWS API request entirely.  But Matthew points out that an asset
increases the chances of the stored data going stale between installer
invocations [1], so with this commit we're just caching in memory.

[1]: https://github.com/openshift/installer/pull/1212#pullrequestreview-204315822